Python analysis checkpoint update

+++ Python_Analysis/Clean_All_2D.py
@@ -39,6 +39,9 @@ for k in range(dimensions.__len__()):
                                       str(budget) + "_" + type_name + "_" + str(cardinality[cc]) + "/"
                     TEMPORAL_RESULT_FOR_BASH = "./temp_result_" + str(dimension) + "D_top" + str(top_k) + "_budget_" + \
                                                str(budget) + "_" + type_name + "_" + str(cardinality[cc]) + "/"
-                    os.system("rm " + TEMPORAL_RESULT + "*")
+                    if not os.path.exists(TEMPORAL_RESULT):
+                        continue
+                    else:
+                        os.system("rm " + TEMPORAL_RESULT + "*")
 
 print("Done .\n")
